Fix return type

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,7 +5,7 @@
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     floor: Function to return the floor of a given float
     @n: float to floor
